[Edit] コードの可読性を向上 [Add] Inventoryを管理するファイル / 各パッケージ を追加 [Other] 変更しすぎて忘れた
